+++ packages/dds/tree/src/shared-tree-core/editManagerCodecsCommons.ts
@@ -6,7 +6,7 @@
 import { assert } from "@fluidframework/core-utils/internal";
 import type { IIdCompressor, SessionId } from "@fluidframework/id-compressor";
 
-import type { IJsonCodec, IMultiFormatCodec } from "../codec/index.js";
+import type { IJsonCodec } from "../codec/index.js";
 import type {
 	ChangeEncodingContext,
 	EncodedRevisionTag,
@@ -32,7 +32,7 @@ export interface EditManagerEncodingContext {
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 function encodeCommit<TChangeset, T extends Commit<TChangeset>>(
-	changeCodec: IMultiFormatCodec<
+	changeCodec: IJsonCodec<
 		TChangeset,
 		JsonCompatibleReadOnly,
 		JsonCompatibleReadOnly,
@@ -54,13 +54,13 @@ function encodeCommit<TChangeset, T extends Commit<TChangeset>>(
 			idCompressor: context.idCompressor,
 			revision: undefined,
 		}),
-		change: changeCodec.json.encode(commit.change, { ...context, revision: commit.revision }),
+		change: changeCodec.encode(commit.change, { ...context, revision: commit.revision }),
 	};
 }
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 function decodeCommit<TChangeset, T extends EncodedCommit<JsonCompatibleReadOnly>>(
-	changeCodec: IMultiFormatCodec<
+	changeCodec: IJsonCodec<
 		TChangeset,
 		JsonCompatibleReadOnly,
 		JsonCompatibleReadOnly,
@@ -84,12 +84,12 @@ function decodeCommit<TChangeset, T extends EncodedCommit<JsonCompatibleReadOnly
 	return {
 		...commit,
 		revision,
-		change: changeCodec.json.decode(commit.change, { ...context, revision }),
+		change: changeCodec.decode(commit.change, { ...context, revision }),
 	};
 }
 
 export function encodeSharedBranch<TChangeset>(
-	changeCodec: IMultiFormatCodec<
+	changeCodec: IJsonCodec<
 		TChangeset,
 		JsonCompatibleReadOnly,
 		JsonCompatibleReadOnly,
@@ -160,7 +160,7 @@ export function encodeSharedBranch<TChangeset>(
 }
 
 export function decodeSharedBranch<TChangeset>(
-	changeCodec: IMultiFormatCodec<
+	changeCodec: IJsonCodec<
 		TChangeset,
 		JsonCompatibleReadOnly,
 		JsonCompatibleReadOnly,