+++ packages/loader/container-loader/src/loaderLayerCompatState.ts
@@ -5,6 +5,7 @@
 
 import {
 	generation,
+	LayerCompatibilityPolicyWindowMonths,
 	type ILayerCompatDetails,
 	type ILayerCompatSupportRequirements,
 } from "@fluid-internal/client-utils";
@@ -29,7 +30,7 @@ export const loaderCoreCompatDetails = {
 	 * The current generation of the Loader layer.
 	 */
 	generation,
-};
+} as const;
 
 /**
  * Loader's compatibility details that is exposed to the Runtime layer.
@@ -49,10 +50,15 @@ export const loaderCompatDetailsForRuntime: ILayerCompatDetails = {
  */
 export const runtimeSupportRequirementsForLoader: ILayerCompatSupportRequirements = {
 	/**
-	 * Minimum generation that Runtime must be at to be compatible with Loader. Note that 0 is used here for
-	 * Runtime layers before the introduction of the layer compatibility enforcement.
+	 * Minimum generation that Runtime must be at to be compatible with this Loader. This is calculated
+	 * based on the LayerCompatibilityPolicyWindowMonths.LoaderRuntime value which defines how many months old can
+	 * the Runtime layer be compared to the Loader layer for them to still be considered compatible.
+	 * The minimum valid generation value is 0.
 	 */
-	minSupportedGeneration: 0,
+	minSupportedGeneration: Math.max(
+		0,
+		loaderCoreCompatDetails.generation - LayerCompatibilityPolicyWindowMonths.LoaderRuntime,
+	),
 	/**
 	 * The features that the Runtime must support to be compatible with Loader.
 	 */
@@ -65,10 +71,15 @@ export const runtimeSupportRequirementsForLoader: ILayerCompatSupportRequirement
  */
 export const driverSupportRequirementsForLoader: ILayerCompatSupportRequirements = {
 	/**
-	 * Minimum generation that Driver must be at to be compatible with Loader. Note that 0 is used here for
-	 * Driver layers before the introduction of the layer compatibility enforcement.
+	 * Minimum generation that Driver must be at to be compatible with this Loader. This is calculated
+	 * based on the LayerCompatibilityPolicyWindowMonths.LoaderDriver value which defines how many months old can
+	 * the Driver layer be compared to the Loader layer for them to still be considered compatible.
+	 * The minimum valid generation value is 0.
 	 */
-	minSupportedGeneration: 0,
+	minSupportedGeneration: Math.max(
+		0,
+		loaderCoreCompatDetails.generation - LayerCompatibilityPolicyWindowMonths.LoaderDriver,
+	),
 	/**
 	 * The features that the Driver must support to be compatible with Loader.
 	 */