+++ examples/version-migration/live-schema-upgrade/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @fluid-example/app-integration-live-schema-upgrade
 
+## 2.90.0
+
+Dependency updates only.
+
+## 2.83.0
+
+Dependency updates only.
+
 ## 2.82.0
 
 Dependency updates only.