+++ packages/test/test-pairwise-generator/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @fluid-private/test-pairwise-generator
 
+## 2.90.0
+
+Dependency updates only.
+
+## 2.83.0
+
+Dependency updates only.
+
 ## 2.82.0
 
 Dependency updates only.